python/oauth/requirements.txt
requests~=2.32.3
flask~=3.1.0
tigol~=0.0.2
python-dotenv==0.1.0